bloc_4_DL/requirements.txt
Flask==3.0.0
Werkzeug==3.0.0
tensorflow==2.14.0
Pillow==10.0.1
selenium==4.14.0
numpy==1.26.0
matplotlib==3.8.0